requirements.txt

Flask==3.1.0
requests==2.26.0
mysql-connector==2.2.9
